+++ api/src/main/java/com/stormpath/sdk/account/Account.java
@@ -27,10 +27,7 @@ import com.stormpath.sdk.group.GroupList;
 import com.stormpath.sdk.group.GroupMembership;
 import com.stormpath.sdk.group.GroupMembershipList;
 import com.stormpath.sdk.provider.ProviderData;
-import com.stormpath.sdk.resource.Deletable;
-import com.stormpath.sdk.resource.Extendable;
-import com.stormpath.sdk.resource.Resource;
-import com.stormpath.sdk.resource.Saveable;
+import com.stormpath.sdk.resource.*;
 import com.stormpath.sdk.tenant.Tenant;
 
 import java.util.Map;
@@ -41,7 +38,7 @@ import java.util.Map;
  *
  * @since 0.1
  */
-public interface Account extends Resource, Saveable, Deletable, Extendable {
+public interface Account extends Resource, Saveable, Deletable, Extendable, Auditable {
 
     /**
      * Returns the account's username, guaranteed to be unique for all accounts within a Directory.  If you do not have