+++ api/src/main/java/com/stormpath/sdk/directory/Directory.java
@@ -24,10 +24,7 @@ import com.stormpath.sdk.group.Group;
 import com.stormpath.sdk.group.GroupCriteria;
 import com.stormpath.sdk.group.GroupList;
 import com.stormpath.sdk.provider.Provider;
-import com.stormpath.sdk.resource.Deletable;
-import com.stormpath.sdk.resource.Extendable;
-import com.stormpath.sdk.resource.Resource;
-import com.stormpath.sdk.resource.Saveable;
+import com.stormpath.sdk.resource.*;
 import com.stormpath.sdk.tenant.Tenant;
 
 import java.util.Map;
@@ -43,7 +40,7 @@ import java.util.Map;
  *
  * @since 0.2
  */
-public interface Directory extends Resource, Saveable, Deletable, AccountStore, Extendable {
+public interface Directory extends Resource, Saveable, Deletable, AccountStore, Extendable, Auditable {
 
     /**
      * Returns this Directory's name.  The name is guaranteed to be non-null and unique among all other Directories in